video-palette-analyzer/requirements.txt
# Core dependencies
numpy>=1.20.0
opencv-python>=4.5.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
Pillow>=8.2.0

# Server dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
flask>=2.0.0
python-multipart>=0.0.5
jinja2>=3.0.0

# Development dependencies
pytest>=6.2.0
black>=21.5b2
isort>=5.9.0